yoro-backend/yoro/requirements.txt
amqp==1.4.6
ansible==1.9.2
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
Django==1.8.2
django-celery==3.1.16
django-cors-headers==1.1.0
django-videothumbs==0.91
djangorestframework==3.1.2
ecdsa==0.13
Jinja2==2.8
kombu==3.0.26
MarkupSafe==0.23
#MySQL-python==1.2.5
paramiko==1.15.2
Pillow==2.8.1
pycrypto==2.6.1
pytz==2015.4
PyYAML==3.11
requests==2.7.0
six==1.9.0
uWSGI==2.0.10
